explicity cast uv_plane_len as int
fixes https://github.com/pupil-labs/pupil/issues/617 - but shows runtime warning:
/Users/wrp/Documents/pupil_labs/code/pupil/pupil_src/shared_modules/av_writer.py:130: VisibleDeprecationWarning: using a non-integer number instead of an integer will result in an error in the future
  y,u,v = input_frame.yuv422

+++ uvc.pyx
@@ -206,7 +206,7 @@ cdef class Frame:
             Y = np.asarray(self._yuv_buffer[:y_plane_len]).reshape(self.height,self.width)
 
             if self.yuv_subsampling == turbojpeg.TJSAMP_422:
-                uv_plane_len = y_plane_len/2
+                uv_plane_len = int(y_plane_len/2)
                 offset = y_plane_len
                 U = np.asarray(self._yuv_buffer[offset:offset+uv_plane_len]).reshape(self.height,self.width/2)
                 offset += uv_plane_len
